Split out traits from base test cases.
