+++ tests/TestCase.php
@@ -1,18 +1,13 @@
 <?php
 
-use Carbon\Carbon;
-use DoSomething\Gateway\Blink;
-use League\OAuth2\Server\CryptKey;
-use Northstar\Auth\Entities\AccessTokenEntity;
-use Northstar\Auth\Entities\ClientEntity;
-use Northstar\Auth\Entities\ScopeEntity;
-use Northstar\Auth\Scope;
-use Northstar\Models\Client;
-use Northstar\Models\User;
-use Northstar\Services\Phoenix;
+use Tests\CreatesApplication;
+use Tests\WithMocks;
+use Tests\WithAuthentication;
 
 abstract class TestCase extends Illuminate\Foundation\Testing\TestCase
 {
+    use CreatesApplication, WithMocks, WithAuthentication;
+
     /**
      * The base URL to use while testing the application.
      *
@@ -26,180 +21,6 @@ abstract class TestCase extends Illuminate\Foundation\Testing\TestCase
      * @var array
      */
     protected $headers = [];
-
-    /**
-     * The Faker generator, for creating test data.
-     *
-     * @var \Faker\Generator
-     */
-    protected $faker;
-
-    /**
-     * The Phoenix API client mock.
-     *
-     * @var \Mockery\MockInterface
-     */
-    protected $phoenixMock;
-
-    /**
-     * The Blink API client mock.
-     *
-     * @var \Mockery\MockInterface
-     */
-    protected $blinkMock;
-
-    /**
-     * Make a new authenticated web user.
-     *
-     * @return \Northstar\Models\User
-     */
-    protected function makeAuthWebUser()
-    {
-        $user = factory(User::class)->create();
-
-        $this->be($user, 'web');
-
-        return $user;
-    }
-
-    /**
-     * Setup the test environment. This is run before *every* single
-     * test method, so avoid doing anything that takes too much time!
-     *
-     * @return void
-     */
-    public function setUp()
-    {
-        parent::setUp();
-
-        $this->serverVariables = $this->transformHeadersToServerVars($this->headers);
-
-        // Get a new Faker generator from Laravel.
-        $this->faker = app(\Faker\Generator::class);
-        $this->faker->addProvider(new FakerPhoneNumber($faker));
-
-        // Reset the testing database & run migrations.
-        app()->make('db')->getMongoDB()->drop();
-        $this->artisan('migrate');
-    }
-
-    /**
-     * Use the given API key for this request.
-     *
-     * @param Client $client
-     * @return $this
-     */
-    public function withLegacyApiKey(Client $client)
-    {
-        $this->serverVariables = array_replace($this->serverVariables, [
-            'HTTP_X-DS-REST-API-Key' => $client->client_secret,
-        ]);
-
-        return $this;
-    }
-
-    /**
-     * Set an API key with the given scopes on the request.
-     *
-     * @param array $scopes
-     * @return $this
-     */
-    public function withLegacyApiKeyScopes(array $scopes)
-    {
-        $client = Client::create([
-            'client_id' => 'testing'.$this->faker->uuid,
-            'scope' => $scopes,
-        ]);
-
-        $this->withLegacyApiKey($client);
-
-        return $this;
-    }
-
-    /**
-     * Make the following request as a normal user with the `user` scope.
-     *
-     * @return $this
-     */
-    public function asNormalUser()
-    {
-        $user = factory(User::class)->create();
-
-        return $this->asUser($user, ['user']);
-    }
-
-    /**
-     * Make the following request as a staff user with the `user` and `role:staff` scopes.
-     *
-     * @return $this
-     */
-    public function asStaffUser()
-    {
-        $staff = factory(User::class, 'staff')->create();
-
-        return $this->asUser($staff, ['user', 'role:staff']);
-    }
-
-    /**
-     * Make the following request as an admin user with the `user` and `role:admin` scopes.
-     *
-     * @return $this
-     */
-    public function asAdminUser()
-    {
-        $admin = factory(User::class, 'admin')->create();
-
-        return $this->asUser($admin, ['user', 'role:admin']);
-    }
-
-    /**
-     * Create a signed JWT to authorize resource requests.
-     *
-     * @param User $user
-     * @param array $scopes
-     * @return $this
-     */
-    public function withAccessToken($scopes = [], $user = null)
-    {
-        $accessToken = new AccessTokenEntity();
-        $accessToken->setClient(new ClientEntity('phpunit', 'PHPUnit', $scopes));
-        $accessToken->setIdentifier(bin2hex(random_bytes(40)));
-        $accessToken->setExpiryDateTime((new \DateTime())->add(new DateInterval('PT1H')));
-
-        if ($user) {
-            $accessToken->setUserIdentifier($user->id);
-            $accessToken->setRole($user->role);
-        }
-
-        foreach ($scopes as $identifier) {
-            if (! array_key_exists($identifier, Scope::all())) {
-                continue;
-            }
-
-            $entity = new ScopeEntity();
-            $entity->setIdentifier($identifier);
-            $accessToken->addScope($entity);
-        }
-
-        $header = 'Bearer '.$accessToken->convertToJWT(new CryptKey(storage_path('keys/private.key'), null, false));
-        $this->serverVariables = array_replace($this->serverVariables, [
-            'HTTP_Authorization' => $header,
-        ]);
-
-        return $this;
-    }
-
-    /**
-     * Create a signed JWT to authorize resource requests.
-     *
-     * @param User $user
-     * @param array $scopes
-     * @return $this
-     */
-    public function asUser($user, $scopes = [])
-    {
-        return $this->withAccessToken($scopes, $user);
-    }
 
     /**
      * Get the raw Mongo document for inspection.
@@ -231,73 +52,6 @@ abstract class TestCase extends Illuminate\Foundation\Testing\TestCase
         ));
 
         return $document;
-    }
-
-    /**
-     * Creates the application.
-     *
-     * @return \Illuminate\Foundation\Application
-     */
-    public function createApplication()
-    {
-        $app = require __DIR__.'/../bootstrap/app.php';
-
-        $app->make('Illuminate\Contracts\Console\Kernel')->bootstrap();
-
-        // Configure a mock for Phoenix & default `createDrupalUser` response.
-        $this->phoenixMock = $this->mock(Phoenix::class);
-        $this->phoenixMock->shouldReceive('createDrupalUser')->andReturnUsing(function () {
-            return $this->faker->unique()->numberBetween(1, 30000000);
-        });
-
-        // Configure a mock for Blink model events.
-        $this->blinkMock = $this->mock(Blink::class);
-        $this->blinkMock->shouldReceive('userCreate')->andReturn(true);
-
-        return $app;
-    }
-
-    /**
-     * Mock a class, and register with the IoC container.
-     *
-     * @param $class String - Class name to mock
-     * @return \Mockery\MockInterface
-     */
-    public function mock($class)
-    {
-        $mock = Mockery::mock($class);
-
-        app()->instance($class, $mock);
-
-        return $mock;
-    }
-
-    /**
-     * Spy on a class.
-     *
-     * @param $class String - Class name to mock
-     * @return \Mockery\MockInterface
-     */
-    public function spy($class)
-    {
-        $spy = Mockery::spy($class);
-
-        app()->instance($class, $spy);
-
-        return $spy;
-    }
-
-    /**
-     * "Freeze" time so we can make assertions based on it.
-     *
-     * @param string $time
-     * @return Carbon
-     */
-    public function mockTime($time = 'now')
-    {
-        Carbon::setTestNow((string) new Carbon($time));
-
-        return Carbon::getTestNow();
     }
 
     /**
